[Refactor] Remove Octave support and simplify to MATLAB-only
Directory Structure Changes:
- Rename scripts/solvers/matlab_octave → scripts/solvers/matlab
- Rename scripts/data_loaders/matlab_octave → scripts/data_loaders/matlab
- Update .gitmodules to reflect new submodule paths

Code Updates:
- Update all Python imports from matlab_octave to matlab
- Fix imports in runner.py, test files, and other Python modules
- Maintain all MATLAB solver functionality (SeDuMi, SDPT3)

Documentation Updates:
- CLAUDE.md: Remove Octave references, update directory structure
- README.md: Focus on MATLAB-only installation and setup
- docs/development/*.md: Update architecture diagrams and references
- Remove all installation instructions for Octave

Testing Results:
- SDP problem arch0: OPTIMAL in 1.597s, objective: -5.665173e-01
- SOCP problem nb: OPTIMAL in 0.919s, objective: -5.070309e-02

Benefits:
- Simplified MVP approach focusing on MATLAB only
- Reduced confusion from dual MATLAB/Octave support
- Cleaner codebase and documentation
- Maintained full MATLAB solver functionality
