+++ scripts/solvers/matlab/setup_matlab_solvers.m
