+++ tests/integration/test_enhanced_matlab_solver.py
@@ -14,7 +14,7 @@ from pathlib import Path
 project_root = Path(__file__).parent.parent.parent
 sys.path.insert(0, str(project_root))
 
-from scripts.solvers.matlab_octave.matlab_runner import MatlabSolver, SeDuMiSolver, SDPT3Solver
+from scripts.solvers.matlab.matlab_runner import MatlabSolver, SeDuMiSolver, SDPT3Solver
 from scripts.data_loaders.problem_loader import ProblemData
 from scripts.utils.logger import get_logger
